added chirps

+++ db_access.py
@@ -1,10 +1,17 @@
 import sqlite3
-import uuid, datetime
+import uuid, datetime, hashlib
 from flask import app, g
 
 from chirpr import app
 
-def add_chirp(text):
+def hash_ps(text):
+    t = hashlib.sha224()
+    t.update(text.encode('utf-8'))
+    t = t.hexdigest()
+    return t
+
+
+def add_chirp(text, uID):
     conn = get_db()
     num = -1
     for c in conn.execute('SELECT id FROM chirp'):
@@ -12,7 +19,7 @@ def add_chirp(text):
     num += 1
     conn.execute('''
         INSERT INTO chirp VALUES (?,?,?,?)
-    ''', (num, text, 1, str(datetime.datetime.utcnow())))
+    ''', (num, text, uID, str(datetime.datetime.utcnow())))
     conn.commit()
     return num
     #c.id, c.body, c.datetime, u.handle FROM chirp c, user u WHERE c.user_id = u.id
@@ -23,13 +30,13 @@ def get_all_chirps():
     return conn.execute('''
         SELECT c.id, c.body, c.datetime, u.handle
         FROM chirp c, user u
-        WHERE c.user_id = u.id
+        WHERE c.user_id = u.id ORDER BY c.datetime
     ''').fetchall()
 
 
-def delete_chirp(chirp_id):
+def delete_chirp(chirp_id, user_id):
     conn = get_db()
-    conn.execute('DELETE FROM chirp WHERE id = :id', {'id': chirp_id})
+    conn.execute('DELETE FROM chirp WHERE id = :id AND user_id = :uid', {'id': chirp_id, 'uid':user_id})
     conn.commit()
 
 
@@ -38,16 +45,25 @@ def get_all_users():
     return conn.execute('SELECT id, handle, admin FROM user').fetchall()
 
 
+def user_for(uid):
+    conn = get_db()
+    for a in conn.execute('SELECT handle FROM user WHERE id=:id', {'id':uid}):
+        return a[0]
+    return ''
+
 def delete_user(user_id):
     conn = get_db()
     conn.execute('DELETE FROM user WHERE id = :id', {'id': user_id})
     conn.commit()
 
 
-def add_user(handle):
+def add_user(handle, password):
+    password = hash_ps(password)
     conn = get_db()
-    conn.execute('INSERT INTO user (handle, admin) values (:handle, :admin)',
-                 {'handle': handle, 'admin': 0})
+    for c in conn.execute('SELECT * FROM user WHERE handle=:handle', {'handle':handle}):
+        return list(c)
+    conn.execute('INSERT INTO user (handle, password, admin) values (:handle, :password, :admin)',
+                 {'handle': handle, 'password':password, 'admin': 0})
     conn.commit()
 
 
@@ -63,6 +79,13 @@ def connect_db():
     return conn
 
 
+def sign_in(handle, password):
+    conn = get_db()
+    password = hash_ps(password)
+    for c in conn.execute('SELECT id FROM user WHERE handle=:handle AND password=:password', {'handle':handle, 'password':password}):
+        print(c)
+        return (True, c[0])
+    return (False, -1)
 @app.teardown_appcontext
 def close_db(error):
     if hasattr(g, 'conn'):
